NEW FEATURE: Task Headers are now visible in the task page.
