+++ app/src/main/java/com/mobdeve/s18/task4today/RecyclerItemTouchHelper.kt
@@ -8,6 +8,7 @@ import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.graphics.drawable.Drawable
 import androidx.core.content.ContextCompat
+import com.mobdeve.s18.task4today.adapter.ToDoAdapter
 import kotlin.Int
 
 // RecyclerItemTouchHelper -> swipe task left to [Edit], right to [Delete]
