--- app/src/main/java/com/mobdeve/s18/task4today/ToDoAdapter.kt
@@ -1,82 +1,0 @@
-package com.mobdeve.s18.task4today
-
-import android.content.Context
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-
-class ToDoAdapter (
-    val activity: MainActivity,
-    private var todoList : ArrayList<TaskModel>,
-    private var dbHelper : DbHelper
-) : RecyclerView.Adapter<MyViewHolder>() {
-
-    fun getContext(): Context{
-        return activity
-    }
-
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) : MyViewHolder {
-        // 1. Create a LayoutInflater
-        val inflater = LayoutInflater.from(parent.context)
-
-        // 2. Inflate a new view
-        val view = inflater.inflate(R.layout.task_layout, parent, false)
-        return MyViewHolder(view)
-    }
-
-    override fun onBindViewHolder(holder: MyViewHolder, position: Int){
-        // db.openDatabase() // don't call this na
-        val item = todoList[position]
-
-        // 3. Bind to-do data to the holder
-        holder.task.text = item.task
-
-        // Temporarily remove the listener before setting the checkbox
-        holder.task.setOnCheckedChangeListener(null) // Remove previous listener
-        holder.task.isChecked = toBoolean(item.status) // Set checkbox
-
-        holder.task.setOnCheckedChangeListener{ _, isChecked ->
-            val status = if(isChecked) 1 else 0
-            dbHelper.updateStatus(item.id, status)
-        }
-
-    }
-
-    private fun toBoolean(n: Int): Boolean {
-        return n != 0
-    }
-
-    override fun getItemCount() : Int {
-        // return the number of to-do items
-        return todoList.size
-    }
-
-    // Set tasks on the screen
-    fun setTasks(todoList: ArrayList<TaskModel>){
-        this.todoList = todoList
-        notifyDataSetChanged()
-    }
-
-    fun editItem(position: Int){
-        val item = todoList[position]
-        val bundle = Bundle().apply{
-            putInt("id", item.id)
-            putString("task", item.task)
-        }
-
-        val fragment = AddNewTask.newInstance(item.id, item.task)
-        fragment.show(activity.supportFragmentManager, AddNewTask.TAG)
-    }
-
-    // Delete task
-    fun deleteItem(position: Int){
-        val item = todoList[position]
-        dbHelper.deleteTask(item.id)
-        todoList.removeAt(position)
-        notifyItemRemoved(position)
-
-    }
-
-}
