+++ app/src/main/java/com/mobdeve/s18/task4today/adapter/HeaderListAdapter.kt
@@ -1,5 +1,6 @@
 package com.mobdeve.s18.task4today.adapter
 
+import android.content.Context
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -8,19 +9,32 @@ import androidx.recyclerview.widget.RecyclerView
 import com.mobdeve.s18.task4today.HeaderModel
 import com.mobdeve.s18.task4today.R
 import android.graphics.Color
+import android.os.Bundle
+import android.widget.ImageView
 import androidx.core.graphics.drawable.DrawableCompat
+import com.mobdeve.s18.task4today.AddNewTask
+import com.mobdeve.s18.task4today.DbHelper
+import com.mobdeve.s18.task4today.MainActivity
+import com.mobdeve.s18.task4today.MyViewHolder
+import com.mobdeve.s18.task4today.TaskModel
 
 // HeaderListAdapter.kt
-class HeaderListAdapter(private val headers: List<HeaderModel>) :
-    RecyclerView.Adapter<HeaderListAdapter.HeaderViewHolder>() {
+class HeaderListAdapter(
+    private val headers: List<HeaderModel>,
+    private val layoutResId: Int // Parameter to allow dynamic layout
+) :
+
+RecyclerView.Adapter<HeaderListAdapter.HeaderViewHolder>() {
 
     inner class HeaderViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-        val titleText = itemView.findViewById<TextView>(R.id.headerTitle)
+        val titleText: TextView = itemView.findViewById(R.id.headerTitle)
+        val addButton: ImageView? = itemView.findViewById(R.id.headerAddButton) // Safe optional
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HeaderViewHolder {
         val view = LayoutInflater.from(parent.context)
-            .inflate(R.layout.format_view_header_list, parent, false)
+            .inflate(layoutResId, parent, false) // The layout is now dynamic
+
         return HeaderViewHolder(view)
     }
 
@@ -39,8 +53,88 @@ class HeaderListAdapter(private val headers: List<HeaderModel>) :
                 // Fallback tint
                 DrawableCompat.setTint(background, Color.GRAY)
             }
+
+            // Optional: add button logic (only if present)
+            holder.addButton?.setOnClickListener {
+                // Add task to this header
+            }
         }
     }
 
     override fun getItemCount(): Int = headers.size
+}
+
+class ToDoAdapter (
+    val activity: MainActivity,
+    private var todoList : ArrayList<TaskModel>,
+    private var dbHelper : DbHelper
+) : RecyclerView.Adapter<MyViewHolder>() {
+
+    fun getContext(): Context {
+        return activity
+    }
+
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) : MyViewHolder {
+        // 1. Create a LayoutInflater
+        val inflater = LayoutInflater.from(parent.context)
+
+        // 2. Inflate a new view
+        val view = inflater.inflate(R.layout.format_task_layout, parent, false)
+        return MyViewHolder(view)
+    }
+
+    override fun onBindViewHolder(holder: MyViewHolder, position: Int){
+        // db.openDatabase() // don't call this na
+        val item = todoList[position]
+
+        // 3. Bind to-do data to the holder
+        holder.task.text = item.task
+
+        // Temporarily remove the listener before setting the checkbox
+        holder.task.setOnCheckedChangeListener(null) // Remove previous listener
+        holder.task.isChecked = toBoolean(item.status) // Set checkbox
+
+        holder.task.setOnCheckedChangeListener{ _, isChecked ->
+            val status = if(isChecked) 1 else 0
+            dbHelper.updateStatus(item.id, status)
+        }
+
+    }
+
+    private fun toBoolean(n: Int): Boolean {
+        return n != 0
+    }
+
+    override fun getItemCount() : Int {
+        // return the number of to-do items
+        return todoList.size
+    }
+
+    // Set tasks on the screen
+    fun setTasks(todoList: ArrayList<TaskModel>){
+        this.todoList = todoList
+        notifyDataSetChanged()
+    }
+
+    fun editItem(position: Int){
+        val item = todoList[position]
+        val bundle = Bundle().apply{
+            putInt("id", item.id)
+            putString("task", item.task)
+        }
+
+        val fragment = AddNewTask.Companion.newInstance(item.id, item.task)
+        fragment.show(activity.supportFragmentManager, AddNewTask.Companion.TAG)
+    }
+
+    // Delete task
+    fun deleteItem(position: Int){
+        val item = todoList[position]
+        dbHelper.deleteTask(item.id)
+        todoList.removeAt(position)
+        notifyItemRemoved(position)
+
+    }
+
 }
